doubao_sample/requirements.txt
pyaudio
websockets
vikingdb-python-sdk
requests
python-dotenv
dataclasses
typing-extensions
sounddevice
numpy
logging
dotenv
mem0ai